Hide sharing assets behind an interface
