+++ modules/features/sharing/src/main/kotlin/au/com/shiftyjelly/pocketcasts/sharing/clip/ShareClipListener.kt
@@ -6,8 +6,8 @@ import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
 import au.com.shiftyjelly.pocketcasts.sharing.SharingClient
 import au.com.shiftyjelly.pocketcasts.sharing.SharingRequest
+import au.com.shiftyjelly.pocketcasts.sharing.ui.BackgroundAssetController
 import au.com.shiftyjelly.pocketcasts.sharing.ui.CardType
-import au.com.shiftyjelly.pocketcasts.sharing.ui.VideoBackgroundController
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.coroutineScope
@@ -18,7 +18,7 @@ internal class ShareClipListener(
     private val fragment: ShareClipFragment,
     private val viewModel: ShareClipViewModel,
     private val sharingClient: SharingClient,
-    private val videoBackgroundController: VideoBackgroundController,
+    private val assetController: BackgroundAssetController,
 ) : ShareClipPageListener {
     override suspend fun onShareClipLink(podcast: Podcast, episode: PodcastEpisode, clipRange: Clip.Range) {
         val request = SharingRequest.clipLink(podcast, episode, clipRange).build()
@@ -40,7 +40,7 @@ internal class ShareClipListener(
     @OptIn(ExperimentalComposeApi::class)
     override suspend fun onShareClipVideo(podcast: Podcast, episode: PodcastEpisode, clipRange: Clip.Range) = coroutineScope {
         launch { delay(1.seconds) } // Launch a delay job to allow the loading animation to run even if clipping happens faster
-        val backgroundImage = videoBackgroundController.create(CardType.Vertical).getOrNull()
+        val backgroundImage = assetController.capture(CardType.Vertical).getOrNull()
         if (backgroundImage == null) {
             Toast.makeText(fragment.requireContext(), "Error", Toast.LENGTH_SHORT).show()
         } else {
